requirements.txt
nltk==3.6.2
pyTelegramBotAPI==3.8.1
python-dotenv==0.18.0
requests==2.25.1
